tests/migrate delta and quote suites to v2
