+++ tests/delta.test.ts
@@ -1,1132 +1,1022 @@
 import * as dotenv from 'dotenv';
-import Web3 from 'web3';
-import { ethers } from 'ethersV5';
-import { ethers as ethersV6 } from 'ethers';
 import fetch from 'isomorphic-unfetch';
 import {
-  constructEthersV5ContractCaller,
-  constructEthersV6ContractCaller,
-  constructFetchFetcher,
   constructPartialSDK,
-  constructWeb3ContractCaller,
+  constructFetchFetcher,
   constructGetDeltaContract,
+  constructGetPartnerFee,
+  constructAllDeltaOrdersHandlers,
+  constructBuildDeltaOrder,
+  constructBuildExternalDeltaOrder,
+  constructBuildTWAPDeltaOrder,
+  constructCancelDeltaOrder,
+  constructGetAgentsList,
+  constructGetBridgeRoutes,
   constructGetDeltaOrders,
   constructGetDeltaPrice,
-  constructBuildDeltaOrder,
-  constructApproveTokenForDelta,
-  constructSignDeltaOrder,
-  constructViemContractCaller,
-  constructGetPartnerFee,
-  SignableDeltaOrderData,
-  DeltaPrice,
+  constructIsTokenSupportedInDelta,
   constructPostDeltaOrder,
+  constructPostExternalDeltaOrder,
+  constructPostTWAPDeltaOrder,
   constructSubmitDeltaOrder,
-  PostDeltaOrderParams,
+  PaginatedResponse,
   FetcherFunction,
-  constructCancelDeltaOrder,
-  constructPreSignDeltaOrder,
-  GetDeltaContractFunctions,
-  constructGetBridgeInfo,
-  BridgeInfo,
 } from '../src';
-import BigNumber from 'bignumber.js';
-
-import erc20abi from './abi/ERC20.json';
-
-import { assert } from 'ts-essentials';
-import { HardhatProvider } from './helpers/hardhat';
-import { privateKeyToAccount } from 'viem/accounts';
-import {
-  Address,
-  createWalletClient,
-  custom,
-  Hash,
-  Hex,
-  publicActions,
-  verifyTypedData,
-} from 'viem';
-import { hardhat } from 'viem/chains';
-import { ZERO_ADDRESS } from '../src/methods/common/orders/buildOrderData';
-import {
-  BridgePriceInfo,
-  DeltaAuctionOrder,
-} from '../src/methods/delta/helpers/types';
+import type { ContractCallerFunctions, TxHash } from '../src/types';
+import type {
+  DeltaAuction,
+  DeltaPrice,
+  DeltaRoute,
+  BuiltDeltaOrder,
+  BridgeRoute,
+} from '../src/methods/delta/types';
 
 dotenv.config();
 
-jest.setTimeout(30 * 1000);
-
+const PARASWAP_DELTA = '0x1111111111111111111111111111111111111111';
 const WETH = '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2';
-const DAI = '0x6B175474E89094C44Da98b954EedeAC495271d0F';
+const DAI = '0x6b175474e89094c44da98b954eedeac495271d0f';
+const USDC_ARB = '0xaf88d065e77c8cc2239327c5edb3a432268e5831';
+const OWNER = '0xac39b311dceb2a4b2f5d8461c1cdaf756f4f7ae9';
+const API_URL = 'https://api.test.invalid';
+const FAKE_SIGNATURE = '0x' + 'ab'.repeat(64);
 
-const chainId = 1;
-const srcToken = WETH;
-const destToken = DAI;
-const srcAmount = (1 * 1e18).toString(); //The source amount multiplied by its decimals
+type FetchSpy = jest.Mock<Promise<any>, Parameters<FetcherFunction>>;
 
-const TEST_MNEMONIC =
-  'radar blur cabbage chef fix engine embark joy scheme fiction master release';
-//0xaC39b311DCEb2A4b2f5d8461c1cdaF756F4F7Ae9
-const wallet = ethers.Wallet.fromMnemonic(TEST_MNEMONIC);
-const walletV6 = ethersV6.HDNodeWallet.fromPhrase(TEST_MNEMONIC);
+function buildPriceFixture(
+  overrides: Partial<DeltaPrice> = {}
+): DeltaPrice {
+  const srcInput = {
+    token: { chainId: 1, address: WETH },
+    amount: '1000000000000000000',
+    amountUSD: '3000',
+  };
+  const destOutput = {
+    token: { chainId: 1, address: DAI },
+    amount: '2950000000000000000000',
+    amountUSD: '2950',
+  };
 
-const web3provider = new Web3(HardhatProvider as any);
-
-const ethersProvider = new ethers.providers.Web3Provider(
-  HardhatProvider as any
-);
-
-const ethersV6Provider = new ethersV6.BrowserProvider(HardhatProvider);
-const signerV6 = walletV6.connect(ethersV6Provider);
-
-const fetchFetcher = constructFetchFetcher(fetch);
-
-const signer = wallet.connect(ethersProvider);
-const senderAddress = signer.address;
-
-const viemWalletClient = createWalletClient({
-  // either walletClient needs to have account set at creation
-  // or provider must own the account (for testing can `await viemTestClient.impersonateAccount({ address: senderAddress });`)
-  // to be able to sign transactions
-  account: privateKeyToAccount(wallet.privateKey as Hex),
-  chain: { ...hardhat, id: chainId },
-  transport: custom(HardhatProvider),
-}).extend(publicActions);
-
-const ethersV5ContractCaller = constructEthersV5ContractCaller(
-  {
-    ethersProviderOrSigner: signer,
-    EthersContract: ethers.Contract,
-  },
-  senderAddress
-);
-
-const ethersV6ContractCaller = constructEthersV6ContractCaller(
-  {
-    ethersV6ProviderOrSigner: signerV6,
-    EthersV6Contract: ethersV6.Contract,
-  },
-  senderAddress
-);
-
-const web3ContractCaller = constructWeb3ContractCaller(
-  web3provider,
-  senderAddress
-);
-
-const viemContractCaller = constructViemContractCaller(
-  viemWalletClient,
-  senderAddress
-);
-
-describe('Delta:methods', () => {
-  const deltaSDK = constructPartialSDK(
-    {
-      chainId: 1,
-      fetcher: fetchFetcher,
-      contractCaller: ethersV5ContractCaller,
-      apiURL: process.env.API_URL,
+  const route: DeltaRoute = {
+    origin: { input: srcInput, output: destOutput },
+    destination: { input: destOutput, output: destOutput },
+    bridge: null,
+    fees: {
+      gas: {
+        token: {
+          chainId: 1,
+          address: '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee',
+        },
+        amount: '500000',
+        amountUSD: '0.5',
+      },
+      bridge: [],
     },
-    constructGetDeltaContract,
-    constructGetDeltaOrders,
-    constructGetDeltaPrice,
-    constructBuildDeltaOrder,
-    constructApproveTokenForDelta,
-    constructGetPartnerFee,
-    constructGetBridgeInfo
-  );
+  };
 
-  describe('Bridge methods', () => {
-    function flattenAllBridgeInfoTokens(bridgeInfo: BridgeInfo): string[] {
-      return Object.values(bridgeInfo).flatMap((chainMap) =>
-        Object.values(chainMap).flat()
-      );
-    }
-    test('Get Bridge Info', async () => {
-      const bridgeInfo = await deltaSDK.getBridgeInfo();
-      expect(Object.keys(bridgeInfo)).toEqual(
-        // allow for more chains to be added in the future
-        expect.arrayContaining(['1', '10', '56', '130', '137', '8453', '42161'])
-      );
+  return {
+    id: 'price-id-1',
+    side: 'SELL',
+    inputToken: { chainId: 1, address: WETH },
+    outputToken: { chainId: 1, address: DAI },
+    route,
+    partner: { name: 'sdk-test', feePercent: 0 },
+    spender: PARASWAP_DELTA,
+    alternatives: [],
+    ...overrides,
+  };
+}
 
-      const defaultNumOfTokens = flattenAllBridgeInfoTokens(bridgeInfo).length;
+function buildCrosschainRoute(): DeltaRoute {
+  const srcInput = {
+    token: { chainId: 1, address: WETH },
+    amount: '1000000000000000000',
+    amountUSD: '3000',
+  };
+  const srcOutputIntermediate = {
+    token: {
+      chainId: 1,
+      address: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
+    },
+    amount: '2980000000',
+    amountUSD: '2980',
+  };
+  const destInput = {
+    token: { chainId: 42161, address: USDC_ARB },
+    amount: '2978000000',
+    amountUSD: '2978',
+  };
+  const destOutput = {
+    token: { chainId: 42161, address: USDC_ARB },
+    amount: '2978000000',
+    amountUSD: '2978',
+  };
 
-      const bridgeInfoDIsallowedBridgeAndSwap = await deltaSDK.getBridgeInfo({
-        allowBridgeAndSwap: false,
-      });
-
-      expect(Object.keys(bridgeInfoDIsallowedBridgeAndSwap)).toEqual(
-        Object.keys(bridgeInfo)
-      );
-      const disallowedNumOfTokens = flattenAllBridgeInfoTokens(
-        bridgeInfoDIsallowedBridgeAndSwap
-      ).length;
-
-      // fewer tokens are available when bridge and swap (swap on destChain after bridge) is not allowed
-      expect(disallowedNumOfTokens).toBeLessThan(defaultNumOfTokens);
-    });
-
-    test('Get Bridge Protocols', async () => {
-      const bridgeProtocols = await deltaSDK.getBridgeProtocols();
-      const expectedToInclude = [
+  return {
+    origin: { input: srcInput, output: srcOutputIntermediate },
+    destination: { input: destInput, output: destOutput },
+    bridge: {
+      protocol: 'Across',
+      tags: ['recommended', 'fastest'],
+      estimatedTimeMs: 90_000,
+      contractParams: {
+        protocolSelector: '0xdeadbeef',
+        outputToken: USDC_ARB,
+        scalingFactor: -12,
+        protocolData: '0xabcd',
+      },
+    },
+    fees: {
+      gas: {
+        token: { chainId: 1, address: WETH },
+        amount: '7000000000000000',
+        amountUSD: '21',
+      },
+      bridge: [
         {
-          displayName: 'Across',
-          icon: expect.any(String),
-          protocol: 'Across',
+          token: { chainId: 42161, address: USDC_ARB },
+          amount: '200000',
+          amountUSD: '0.2',
         },
-        {
-          displayName: 'Relay',
-          icon: expect.any(String),
-          protocol: 'Relay',
-        },
-      ];
-      expect(bridgeProtocols).toEqual(
-        // allow for more bridges to be added in the future
-        expect.arrayContaining(expectedToInclude)
-      );
-    });
-  });
+      ],
+    },
+  };
+}
 
-  test('Get Delta Price', async () => {
-    const deltaPrice = await deltaSDK.getDeltaPrice({
-      srcToken: srcToken,
-      destToken: destToken,
-      amount: srcAmount,
-      userAddress: senderAddress,
-      srcDecimals: 18,
-      destDecimals: 18,
-    });
-
-    const staticDeltaPrice: typeof deltaPrice = {
-      ...deltaPrice,
-      partnerFee: NaN, // dynamic number
-      destAmount: 'dynamic_number',
-      destAmountBeforeFee: 'dynamic_number',
-      srcUSD: 'dynamic_number',
-      destUSD: 'dynamic_number',
-      destUSDBeforeFee: 'dynamic_number',
-      receivedDestAmount: 'dynamic_number',
-      receivedDestUSD: 'dynamic_number',
-      receivedDestAmountBeforeFee: 'dynamic_number',
-      receivedDestUSDBeforeFee: 'dynamic_number',
-      gasCost: 'dynamic_number',
-      gasCostBeforeFee: 'dynamic_number',
-      gasCostUSD: 'dynamic_number',
-      gasCostUSDBeforeFee: 'dynamic_number',
-      hmac: 'dynamic_string',
-    };
-
-    expect(staticDeltaPrice).toMatchSnapshot();
-  });
-
-  describe('Get Delta Price Crosschain', () => {
-    const destChainId = 10;
-    const ETH = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE';
-    const WETH = '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2';
-    const WETH_ON_OPTIMISM = '0x4200000000000000000000000000000000000006';
-    // const DAI_TOKEN_ON_ETHEREUM = '0x6B175474E89094C44Da98b954EedeAC495271d0F';
-    const DAI_TOKEN_ON_OPTIMISM = '0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1';
-
-    test('Get Delta Price Crosschain/destToken=random', async () => {
-      const deltaPrice = await deltaSDK.getDeltaPrice({
-        srcToken: WETH,
-        destToken: DAI_TOKEN_ON_OPTIMISM,
-        amount: srcAmount,
-        userAddress: senderAddress,
-        srcDecimals: 18,
-        destDecimals: 18,
-        destChainId,
-      });
-
-      const staticDeltaPrice: typeof deltaPrice = {
-        ...deltaPrice,
-        partnerFee: NaN, // dynamic number
-        destToken: 'dynamic_address', // will no longer match DAI_TOKEN_ON_ETHEREUM if bridge is Relay
-        bridge: {
-          ...deltaPrice.bridge,
-          protocolData: 'dynamic_string',
-          protocolSelector: 'dynamic_string',
-          scalingFactor: NaN, // dynamic number
-        },
-        bridgeInfo: {
-          ...deltaPrice.bridgeInfo,
-          bestReturn: true,
-          destAmountAfterBridge: 'dynamic_number',
-          destUSDAfterBridge: 'dynamic_number',
-          estimatedTimeMs: NaN, // dynamic number
-          fastest: true,
-          fees: [], // dynamic array
-          protocolName: 'dynamic_string' as BridgePriceInfo['protocolName'],
-          recommended: true,
-        },
-        destAmount: 'dynamic_number',
-        destAmountBeforeFee: 'dynamic_number',
-        srcUSD: 'dynamic_number',
-        destUSD: 'dynamic_number',
-        destUSDBeforeFee: 'dynamic_number',
-        receivedDestAmount: 'dynamic_number',
-        receivedDestUSD: 'dynamic_number',
-        receivedDestAmountBeforeFee: 'dynamic_number',
-        receivedDestUSDBeforeFee: 'dynamic_number',
-        gasCost: 'dynamic_number',
-        gasCostBeforeFee: 'dynamic_number',
-        gasCostUSD: 'dynamic_number',
-        gasCostUSDBeforeFee: 'dynamic_number',
-        hmac: 'dynamic_string',
-        availableBridges: [], // dynamic array of bridge variants
-      };
-
-      // will no longer be true if swap is through Relay
-      // expect(deltaPrice.destToken).toEqual(DAI_TOKEN_ON_ETHEREUM.toLowerCase());
-      expect(staticDeltaPrice).toMatchSnapshot();
-      expect(deltaPrice.bridge.destinationChainId).toEqual(destChainId);
-      expect(deltaPrice.bridge.outputToken).toEqual(
-        DAI_TOKEN_ON_OPTIMISM.toLowerCase()
-      );
-    });
-
-    test('Get Delta Price Crosschain/destToken=WETH', async () => {
-      const deltaPrice = await deltaSDK.getDeltaPrice({
-        srcToken: WETH,
-        destToken: WETH_ON_OPTIMISM,
-        amount: srcAmount,
-        userAddress: senderAddress,
-        srcDecimals: 18,
-        destDecimals: 18,
-        destChainId,
-      });
-
-      const staticDeltaPrice: typeof deltaPrice = {
-        ...deltaPrice,
-        partnerFee: NaN, // dynamic number
-        bridge: {
-          ...deltaPrice.bridge,
-          protocolData: 'dynamic_string',
-          protocolSelector: 'dynamic_string',
-          scalingFactor: NaN, // dynamic number
-        },
-        bridgeInfo: {
-          ...deltaPrice.bridgeInfo,
-          bestReturn: true,
-          destAmountAfterBridge: 'dynamic_number',
-          destUSDAfterBridge: 'dynamic_number',
-          estimatedTimeMs: NaN, // dynamic number
-          fastest: true,
-          fees: [], // dynamic array
-          protocolName: 'dynamic_string' as BridgePriceInfo['protocolName'],
-          recommended: true,
-        },
-        destAmount: 'dynamic_number',
-        destAmountBeforeFee: 'dynamic_number',
-        srcUSD: 'dynamic_number',
-        destUSD: 'dynamic_number',
-        destUSDBeforeFee: 'dynamic_number',
-        receivedDestAmount: 'dynamic_number',
-        receivedDestUSD: 'dynamic_number',
-        receivedDestAmountBeforeFee: 'dynamic_number',
-        receivedDestUSDBeforeFee: 'dynamic_number',
-        gasCost: 'dynamic_number',
-        gasCostBeforeFee: 'dynamic_number',
-        gasCostUSD: 'dynamic_number',
-        gasCostUSDBeforeFee: 'dynamic_number',
-        hmac: 'dynamic_string',
-        availableBridges: [], // dynamic array of bridge variants
-      };
-
-      expect(staticDeltaPrice).toMatchSnapshot();
-      expect(deltaPrice.bridge.destinationChainId).toEqual(destChainId);
-      expect(deltaPrice.bridge.outputToken).toEqual(
-        WETH_ON_OPTIMISM.toLowerCase()
-      );
-    });
-
-    test('Get Delta Price Crosschain/destToken=ETH', async () => {
-      const deltaPrice = await deltaSDK.getDeltaPrice({
-        srcToken: WETH,
-        destToken: ETH,
-        amount: srcAmount,
-        userAddress: senderAddress,
-        srcDecimals: 18,
-        destDecimals: 18,
-        destChainId,
-      });
-
-      const staticDeltaPrice: typeof deltaPrice = {
-        ...deltaPrice,
-        partnerFee: NaN, // dynamic number
-        bridge: {
-          ...deltaPrice.bridge,
-          outputToken: 'dynamic_hash', // WETH or ETH depending on bridge used
-          protocolData: 'dynamic_string',
-          protocolSelector: 'dynamic_string',
-          scalingFactor: NaN, // dynamic number
-        },
-        bridgeInfo: {
-          ...deltaPrice.bridgeInfo,
-          bestReturn: true,
-          destAmountAfterBridge: 'dynamic_number',
-          destUSDAfterBridge: 'dynamic_number',
-          estimatedTimeMs: NaN, // dynamic number
-          fastest: true,
-          fees: [], // dynamic array
-          protocolName: 'dynamic_string' as BridgePriceInfo['protocolName'],
-          recommended: true,
-        },
-        destAmount: 'dynamic_number',
-        destAmountBeforeFee: 'dynamic_number',
-        srcUSD: 'dynamic_number',
-        destUSD: 'dynamic_number',
-        destUSDBeforeFee: 'dynamic_number',
-        receivedDestAmount: 'dynamic_number',
-        receivedDestUSD: 'dynamic_number',
-        receivedDestAmountBeforeFee: 'dynamic_number',
-        receivedDestUSDBeforeFee: 'dynamic_number',
-        gasCost: 'dynamic_number',
-        gasCostBeforeFee: 'dynamic_number',
-        gasCostUSD: 'dynamic_number',
-        gasCostUSDBeforeFee: 'dynamic_number',
-        hmac: 'dynamic_string',
-        availableBridges: [], // dynamic array of bridge variants
-      };
-
-      expect(staticDeltaPrice).toMatchSnapshot();
-      expect(deltaPrice.bridge.destinationChainId).toEqual(destChainId);
-      // bridge.outputToken = WETH|ETH for destToken=ETH|WETH on destChain depending on bridge used;
-      // wrap/unwrap logic is determined by bridge.multiCallHandler presence
-      expect([WETH_ON_OPTIMISM.toLowerCase(), ETH.toLowerCase()]).toContain(
-        deltaPrice.bridge.outputToken
-      );
-    });
-  });
-
-  test('Get Delta Contract', async () => {
-    const deltaContract = await deltaSDK.getDeltaContract();
-    expect(deltaContract).toMatchInlineSnapshot(
-      `"0x0000000000bbf5c5fd284e657f01bd000933c96d"`
-    );
-  });
-
-  test('Approve Token For Delta', async () => {
-    const deltaContract = await deltaSDK.getDeltaContract();
-    assert(deltaContract, 'Delta contract not found');
-
-    const allowanceBefore = await getTokenAllowance({
-      tokenAddress: DAI,
-      owner: senderAddress,
-      spender: deltaContract,
-    });
-
-    expect(allowanceBefore.toString()).toEqual('0');
-
-    const amount = '1000000000000000000'; // 1 DAI
-    const tx = await deltaSDK.approveTokenForDelta(amount, DAI);
-    expect(tx).toBeDefined();
-    await tx.wait();
-
-    const allowanceAfter = await getTokenAllowance({
-      tokenAddress: DAI,
-      owner: senderAddress,
-      spender: deltaContract,
-    });
-
-    expect(allowanceAfter.toString()).toEqual(amount);
-  });
-
-  test('Get Delta Orders for user', async () => {
-    const userWithOrders = '0x76176C2971300217E9f48E3dD4e40591500b96Ff';
-
-    const deltaOrders = await deltaSDK.getDeltaOrders({
-      userAddress: userWithOrders,
-    });
-
-    // Orders that we know the user had in the past
-    const staticSliceOfPastOrders = deltaOrders.slice(-2); // first 2 orders historically
-    expect(staticSliceOfPastOrders).toMatchSnapshot();
-  });
-
-  test('Get Delta Order by Id and Hash', async () => {
-    const orderId = '7ec0dc82-98ad-4501-9f46-03e31e51098f';
-    const deltaOrder = await deltaSDK.getDeltaOrderById(orderId);
-    expect(deltaOrder).toMatchSnapshot();
-    expect(deltaOrder).toBeDefined();
-    assert(
-      deltaOrder?.orderHash,
-      "Delta order not found or doesn't have orderHash"
-    );
-
-    const orderByHash = await deltaSDK.getDeltaOrderByHash(
-      deltaOrder.orderHash
-    );
-    expect(orderByHash).toEqual(deltaOrder);
-  });
-
-  test('Get PartnerFee', async () => {
-    const partnerFee = await deltaSDK.getPartnerFee({ partner: 'paraswap.io' });
-    expect(partnerFee).toMatchInlineSnapshot(`
-      {
-        "partnerAddress": "0xc85f5d432b7fa25287c7e0cb88139a1a4c37f565",
-        "partnerFee": 0.15,
-        "takeSurplus": false,
-      }
-    `);
-  });
-
-  test('Build Delta Order', async () => {
-    const sampleDeltaPrice: DeltaPrice = {
-      destAmount: '3163263721766488892666',
-      destAmountBeforeFee: '3194635547945152526200',
-      receivedDestAmount: '3163263721766488892666',
-      destToken: '0x6b175474e89094c44da98b954eedeac495271d0f',
-      destUSD: '3166.4269854931',
-      receivedDestUSD: '3166.4269854931',
-      destUSDBeforeFee: '3197.8301834931',
-      gasCost: '347788',
-      gasCostBeforeFee: '124240',
-      gasCostUSD: '31.403198',
-      gasCostUSDBeforeFee: '11.218137',
-      partner: 'anon',
-      partnerFee: 0,
-      srcAmount: '1000000000000000000',
-      srcToken: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-      srcUSD: '3191.5500000000',
-      hmac: '1234aeb',
-      bridge: {
-        protocolSelector: '0x00000000',
-        destinationChainId: 0,
-        outputToken: ZERO_ADDRESS,
-        scalingFactor: 0,
-        protocolData: '0x',
-      },
-    };
-
-    const slippagePercent = 0.5;
-    const destAmountAfterSlippage = decreaseBySlippage(
-      sampleDeltaPrice.destAmount,
-      slippagePercent
-    );
-
-    const amount = '1000000000000000000'; // 1 DAI
-
-    const signableOrderData = await deltaSDK.buildDeltaOrder({
-      deltaPrice: sampleDeltaPrice,
-      owner: senderAddress,
-      // beneficiary: anotherAccount, // if need to send destToken to another account
-      // permit: "0x1234...", // if signed a Permit1 or Permit2 TransferFrom for DeltaContract
-      srcToken: WETH,
-      destToken: DAI,
-      srcAmount: amount,
-      destAmount: destAmountAfterSlippage, // minimum acceptable destAmount
-      partnerAddress: ZERO_ADDRESS,
-    });
-
-    const staticSignableOrderData: typeof signableOrderData = {
-      ...signableOrderData,
-      data: {
-        ...signableOrderData.data,
-        deadline: NaN, // dynamic number
-        nonce: 'dynamic_number',
-      },
-    };
-    //                             for ZERO partnerAddress capSurplus (true) shifted (<< 9) = 512
-    expect(signableOrderData.data.partnerAndFee).toEqual((1 << 9).toString());
-    expect(staticSignableOrderData).toMatchSnapshot();
-  });
-
-  test('Build Delta Order with slippage (SELL)', async () => {
-    const sampleDeltaPrice: DeltaPrice = {
-      destAmount: '3163263721766488892666',
-      destAmountBeforeFee: '3194635547945152526200',
-      receivedDestAmount: '3163263721766488892666',
-      destToken: '0x6b175474e89094c44da98b954eedeac495271d0f',
-      destUSD: '3166.4269854931',
-      receivedDestUSD: '3166.4269854931',
-      destUSDBeforeFee: '3197.8301834931',
-      gasCost: '347788',
-      gasCostBeforeFee: '124240',
-      gasCostUSD: '31.403198',
-      gasCostUSDBeforeFee: '11.218137',
-      partner: 'anon',
-      partnerFee: 0,
-      srcAmount: '1000000000000000000',
-      srcToken: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-      srcUSD: '3191.5500000000',
-      hmac: '1234aeb',
-      bridge: {
-        protocolSelector: '0x00000000',
-        destinationChainId: 0,
-        outputToken: ZERO_ADDRESS,
-        scalingFactor: 0,
-        protocolData: '0x',
-      },
-    };
-
-    const slippageBps = 50; // 50 bps = 0.5%
-    const BPS_BASE = BigInt(10_000);
-    const expectedDestAmount = (
-      (BigInt(sampleDeltaPrice.destAmount) * (BPS_BASE - BigInt(slippageBps))) /
-      BPS_BASE
-    ).toString(10);
-
-    const amount = '1000000000000000000';
-
-    const signableOrderData = await deltaSDK.buildDeltaOrder({
-      deltaPrice: sampleDeltaPrice,
-      owner: senderAddress,
-      srcToken: WETH,
-      destToken: DAI,
-      srcAmount: amount,
-      slippage: slippageBps,
-      partnerAddress: ZERO_ADDRESS,
-    });
-
-    expect(signableOrderData.data.srcAmount).toEqual(amount);
-    expect(signableOrderData.data.destAmount).toEqual(expectedDestAmount);
-  });
-
-  test('Build Delta Order with slippage (BUY)', async () => {
-    const sampleDeltaPrice: DeltaPrice = {
-      destAmount: '3163263721766488892666',
-      destAmountBeforeFee: '3194635547945152526200',
-      receivedDestAmount: '3163263721766488892666',
-      destToken: '0x6b175474e89094c44da98b954eedeac495271d0f',
-      destUSD: '3166.4269854931',
-      receivedDestUSD: '3166.4269854931',
-      destUSDBeforeFee: '3197.8301834931',
-      gasCost: '347788',
-      gasCostBeforeFee: '124240',
-      gasCostUSD: '31.403198',
-      gasCostUSDBeforeFee: '11.218137',
-      partner: 'anon',
-      partnerFee: 0,
-      srcAmount: '1000000000000000000',
-      srcToken: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-      srcUSD: '3191.5500000000',
-      hmac: '1234aeb',
-      bridge: {
-        protocolSelector: '0x00000000',
-        destinationChainId: 0,
-        outputToken: ZERO_ADDRESS,
-        scalingFactor: 0,
-        protocolData: '0x',
-      },
-    };
-
-    const slippageBps = 50; // 50 bps = 0.5%
-    const destAmount = '3163263721766488892666';
-
-    const BPS_BASE = BigInt(10_000);
-    const expectedSrcAmount = (
-      (BigInt(sampleDeltaPrice.srcAmount) * (BPS_BASE + BigInt(slippageBps))) /
-      BPS_BASE
-    ).toString(10);
-
-    const signableOrderData = await deltaSDK.buildDeltaOrder({
-      deltaPrice: sampleDeltaPrice,
-      owner: senderAddress,
-      srcToken: WETH,
-      destToken: DAI,
-      destAmount,
-      slippage: slippageBps,
-      side: 'BUY',
-      partnerAddress: ZERO_ADDRESS,
-    });
-
-    expect(signableOrderData.data.destAmount).toEqual(destAmount);
-    expect(signableOrderData.data.srcAmount).toEqual(expectedSrcAmount);
-  });
-
-  let signature = '';
-
-  test.each([
-    ['ethersV5', ethersV5ContractCaller],
-    ['ethersV6', ethersV6ContractCaller],
-    ['web3', web3ContractCaller],
-    ['viem', viemContractCaller],
-  ])('sign Delta Order with %s', async (_libName, contractCaller) => {
-    const sdk = constructPartialSDK(
-      {
+/** Minimal BuiltDeltaOrder fixture representing a server-built order. */
+function buildBuiltOrderFixture(
+  value: Record<string, unknown> = {},
+  orderHash = '0xdeadbeef1234'
+): BuiltDeltaOrder {
+  return {
+    toSign: {
+      domain: {
+        name: 'Portikus',
+        version: '2.0.0',
         chainId: 1,
-        fetcher: fetchFetcher,
-        contractCaller,
-        apiURL: process.env.API_URL,
+        verifyingContract: PARASWAP_DELTA,
       },
-      constructSignDeltaOrder
-    );
-
-    const sampleOrder: SignableDeltaOrderData = {
-      data: {
-        beneficiary: '0xaC39b311DCEb2A4b2f5d8461c1cdaF756F4F7Ae9',
-        deadline: 1731328853,
-        destAmount: '3147447403157656698880',
-        destToken: '0x6B175474E89094C44Da98b954EedeAC495271d0F',
-        expectedAmount: '3163263721766488892666',
-        nonce: '1731325253703',
-        owner: '0xaC39b311DCEb2A4b2f5d8461c1cdaF756F4F7Ae9',
-        partnerAndFee: '0',
-        permit: '0x',
+      types: {
+        Order: [
+          { name: 'owner', type: 'address' },
+          { name: 'beneficiary', type: 'address' },
+          { name: 'srcToken', type: 'address' },
+          { name: 'destToken', type: 'address' },
+          { name: 'srcAmount', type: 'uint256' },
+          { name: 'destAmount', type: 'uint256' },
+          { name: 'expectedAmount', type: 'uint256' },
+          { name: 'deadline', type: 'uint256' },
+          { name: 'kind', type: 'uint8' },
+          { name: 'nonce', type: 'uint256' },
+          { name: 'partnerAndFee', type: 'uint256' },
+          { name: 'permit', type: 'bytes' },
+          { name: 'metadata', type: 'bytes' },
+          { name: 'bridge', type: 'Bridge' },
+        ],
+        Bridge: [
+          { name: 'protocolSelector', type: 'bytes4' },
+          { name: 'destinationChainId', type: 'uint256' },
+          { name: 'outputToken', type: 'address' },
+          { name: 'scalingFactor', type: 'int8' },
+          { name: 'protocolData', type: 'bytes' },
+        ],
+      },
+      value: {
+        owner: OWNER,
+        beneficiary: OWNER,
+        srcToken: WETH,
+        destToken: DAI,
         srcAmount: '1000000000000000000',
-        srcToken: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
+        destAmount: '2950000000000000000000',
+        expectedAmount: '2950000000000000000000',
+        deadline: 9999999999,
         kind: 0,
+        nonce: '12345',
+        permit: '0x',
+        partnerAndFee: '0',
         metadata: '0x',
         bridge: {
           protocolSelector: '0x00000000',
           destinationChainId: 0,
-          outputToken: ZERO_ADDRESS,
+          outputToken: '0x0000000000000000000000000000000000000000',
           scalingFactor: 0,
           protocolData: '0x',
         },
+        ...value,
       },
-      domain: {
-        chainId: 1,
-        name: 'Portikus',
-        verifyingContract: '0x0000000000bbf5c5fd284e657f01bd000933c96d',
-        version: '2.0.0',
-      },
-      types: {
-        Order: [
-          {
-            name: 'owner',
-            type: 'address',
-          },
-          {
-            name: 'beneficiary',
-            type: 'address',
-          },
-          {
-            name: 'srcToken',
-            type: 'address',
-          },
-          {
-            name: 'destToken',
-            type: 'address',
-          },
-          {
-            name: 'srcAmount',
-            type: 'uint256',
-          },
-          {
-            name: 'destAmount',
-            type: 'uint256',
-          },
-          {
-            name: 'expectedAmount',
-            type: 'uint256',
-          },
-          {
-            name: 'deadline',
-            type: 'uint256',
-          },
-          {
-            name: 'kind',
-            type: 'uint8',
-          },
-          {
-            name: 'nonce',
-            type: 'uint256',
-          },
-          {
-            name: 'partnerAndFee',
-            type: 'uint256',
-          },
-          {
-            name: 'permit',
-            type: 'bytes',
-          },
-          {
-            name: 'metadata',
-            type: 'bytes',
-          },
-          {
-            name: 'bridge',
-            type: 'Bridge',
-          },
-        ],
-        Bridge: [
-          {
-            name: 'protocolSelector',
-            type: 'bytes4',
-          },
-          {
-            name: 'destinationChainId',
-            type: 'uint256',
-          },
-          {
-            name: 'outputToken',
-            type: 'address',
-          },
-          {
-            name: 'scalingFactor',
-            type: 'int8',
-          },
-          {
-            name: 'protocolData',
-            type: 'bytes',
-          },
-        ],
-      },
-    };
-
-    const deltaOrderSignature = await sdk.signDeltaOrder(sampleOrder);
-    if (!signature) signature = deltaOrderSignature;
-    // signatures match between libraries
-    expect(deltaOrderSignature).toEqual(signature);
-  });
-
-  let cancelSignature = '';
-
-  const sampleOrderId = '7ec0dc82-98ad-4501-9f46-03e31e51098f';
-
-  test.each([
-    ['ethersV5', ethersV5ContractCaller],
-    ['ethersV6', ethersV6ContractCaller],
-    ['web3', web3ContractCaller],
-    ['viem', viemContractCaller],
-  ])(
-    'sign Cancel Delta Order Request with %s',
-    async (_libName, contractCaller) => {
-      const sdk = constructPartialSDK(
-        {
-          chainId: 1,
-          fetcher: fetchFetcher,
-          contractCaller,
-          apiURL: process.env.API_URL,
-        },
-        constructCancelDeltaOrder
-      );
-
-      const deltaCancelSignature = await sdk.signCancelLimitDeltaOrderRequest({
-        orderIds: [sampleOrderId],
-      });
-
-      const valid = await verifySignedCancelRequest({
-        orderId: sampleOrderId,
-        signature: deltaCancelSignature,
-        address: senderAddress,
-        chainId: sdk.chainId,
-      });
-
-      expect(valid).toBe(true);
-
-      if (!cancelSignature) cancelSignature = deltaCancelSignature;
-      // signatures match between libraries
-      expect(deltaCancelSignature).toEqual(cancelSignature);
-    }
-  );
-
-  const dummyFetcher: FetcherFunction = (params) => {
-    // intercept POST requests
-    if (params.method === 'POST') {
-      return params as any;
-    }
-
-    return fetchFetcher(params);
-  };
-
-  const mockFetch = jest.fn(dummyFetcher);
-
-  const dummySDK = constructPartialSDK(
-    {
-      chainId: 1,
-      fetcher: mockFetch as FetcherFunction,
-      contractCaller: ethersV5ContractCaller,
-      apiURL: process.env.API_URL,
     },
-    constructPostDeltaOrder,
-    constructSubmitDeltaOrder
-  );
+    orderHash,
+  };
+}
 
-  test('Post Delta Order', async () => {
-    const sampleOrderData: SignableDeltaOrderData['data'] = {
-      beneficiary: '0xaC39b311DCEb2A4b2f5d8461c1cdaF756F4F7Ae9',
-      deadline: NaN, // dynamic number
-      destAmount: '3147447403157656698880',
-      destToken: '0x6B175474E89094C44Da98b954EedeAC495271d0F',
-      expectedAmount: '3163263721766488892666',
-      nonce: 'dynamic_number',
-      owner: '0xaC39b311DCEb2A4b2f5d8461c1cdaF756F4F7Ae9',
-      partnerAndFee: '0',
-      permit: '0x',
-      srcAmount: '1000000000000000000',
-      srcToken: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-      kind: 0,
-      metadata: '0x',
-      bridge: {
-        protocolSelector: '0x00000000',
-        destinationChainId: 0,
-        outputToken: ZERO_ADDRESS,
-        scalingFactor: 0,
-        protocolData: '0x',
-      },
-    };
+function makeMockContractCaller(): ContractCallerFunctions<TxHash> {
+  return {
+    staticCall: jest.fn(async () => {
+      throw new Error(
+        'staticCall should not be invoked in v2 fetch-only tests'
+      );
+    }),
+    transactCall: jest.fn(async () => '0xfeedface' as TxHash),
+    signTypedDataCall: jest.fn(async () => FAKE_SIGNATURE),
+  };
+}
 
-    const sampleSignature = '0x1234....';
+function makeFetcher(handler: (params: any) => any): FetchSpy {
+  return jest.fn(async (params) => handler(params)) as FetchSpy;
+}
 
-    const input = {
-      order: sampleOrderData,
-      signature: sampleSignature,
-    };
-
-    await dummySDK.postDeltaOrder(input);
-
-    expect(mockFetch).toHaveBeenLastCalledWith({
-      data: { ...input, chainId: dummySDK.chainId },
-      method: 'POST',
-      url: `${dummySDK.apiURL}/delta/orders`,
+describe('Delta: fetch methods', () => {
+  test('getDeltaPrice hits /delta/v2/prices and returns DeltaPrice', async () => {
+    const fixture = buildPriceFixture();
+    const fetcher = makeFetcher(({ url, method }) => {
+      expect(method).toBe('GET');
+      expect(url.startsWith(`${API_URL}/delta/v2/prices?`)).toBe(true);
+      expect(url).toContain(`srcToken=${WETH}`);
+      expect(url).toContain(`destToken=${DAI}`);
+      expect(url).toContain('amount=1000000000000000000');
+      expect(url).toContain('chainId=1');
+      expect(url).toContain('side=SELL');
+      return fixture;
     });
-  });
 
-  test('Submit(=build+sign+post) Delta Order', async () => {
-    const sampleDeltaPrice: DeltaPrice = {
-      destAmount: '3163263721766488892666',
-      destAmountBeforeFee: '3194635547945152526200',
-      receivedDestAmount: '3163263721766488892666',
-      destToken: '0x6b175474e89094c44da98b954eedeac495271d0f',
-      destUSD: '3166.4269854931',
-      receivedDestUSD: '3166.4269854931',
-      destUSDBeforeFee: '3197.8301834931',
-      gasCost: '347788',
-      gasCostBeforeFee: '124240',
-      gasCostUSD: '31.403198',
-      gasCostUSDBeforeFee: '11.218137',
-      partner: 'anon',
-      partnerFee: 0,
-      srcAmount: '1000000000000000000',
-      srcToken: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-      srcUSD: '3191.5500000000',
-      hmac: '1234aeb',
-      bridge: {
-        protocolSelector: '0x00000000',
-        destinationChainId: 0,
-        outputToken: ZERO_ADDRESS,
-        scalingFactor: 0,
-        protocolData: '0x',
-      },
-    };
+    const { getDeltaPrice } = constructGetDeltaPrice({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
 
-    const slippagePercent = 0.5;
-    const destAmountAfterSlippage = decreaseBySlippage(
-      sampleDeltaPrice.destAmount,
-      slippagePercent
-    );
-
-    const amount = '1000000000000000000'; // 1 DAI
-
-    const input = {
-      deltaPrice: sampleDeltaPrice,
-      owner: senderAddress,
-      // beneficiary: anotherAccount, // if need to send destToken to another account
-      // permit: "0x1234...", // if signed a Permit1 or Permit2 TransferFrom for DeltaContract
+    const price = await getDeltaPrice({
       srcToken: WETH,
       destToken: DAI,
-      srcAmount: amount,
-      destAmount: destAmountAfterSlippage, // minimum acceptable destAmount
-      partnerAddress: ZERO_ADDRESS,
-    };
+      amount: '1000000000000000000',
+      srcDecimals: 18,
+      destDecimals: 18,
+    });
 
-    await dummySDK.submitDeltaOrder(input);
-
-    const callArgs = mockFetch.mock.lastCall?.[0];
-    assert(callArgs, 'No fetch call was made');
-    assert('data' in callArgs, 'No data was sent in the fetch call');
-    const { order, signature } = callArgs.data as PostDeltaOrderParams;
-
-    expect(signature).toBeDefined();
-
-    const staticSignedOrderData: SignableDeltaOrderData['data'] = {
-      ...order,
-      deadline: NaN, // dynamic number
-      nonce: 'dynamic_number',
-    };
-
-    //                             for ZERO partnerAddress capSurplus (true) shifted (<< 9) = 512
-    expect(order.partnerAndFee).toEqual((1 << 9).toString());
-    expect(staticSignedOrderData).toMatchSnapshot();
+    expect(price).toEqual(fixture);
+    expect(price.route.origin.input.amount).toBe('1000000000000000000');
+    expect(price.route.bridge).toBeNull();
+    expect(fetcher).toHaveBeenCalledTimes(1);
   });
 
-  describe('PreSign Delta Order', () => {
-    const sdk = constructPartialSDK(
-      {
-        chainId: 1,
-        fetcher: fetchFetcher,
-        contractCaller: viemContractCaller,
-        apiURL: process.env.API_URL,
-      },
-      constructPreSignDeltaOrder,
-      constructGetDeltaContract
-    );
+  test('getDeltaPrice passes destChainId for cross-chain', async () => {
+    const fixture = buildPriceFixture({
+      route: buildCrosschainRoute(),
+      outputToken: { chainId: 42161, address: USDC_ARB },
+    });
+    const fetcher = makeFetcher(({ url }) => {
+      expect(url).toContain('destChainId=42161');
+      return fixture;
+    });
 
-    const sampleOrder: DeltaAuctionOrder = {
-      owner: '0xaC39b311DCEb2A4b2f5d8461c1cdaF756F4F7Ae9',
-      beneficiary: '0xaC39b311DCEb2A4b2f5d8461c1cdaF756F4F7Ae9',
-      srcToken: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-      destToken: '0x6B175474E89094C44Da98b954EedeAC495271d0F',
-      srcAmount: '1000000000000000000',
-      destAmount: '3147447403157656698880',
-      expectedAmount: '3163263721766488892666',
-      kind: 0,
-      metadata: '0x',
-      deadline: 1731328853,
-      nonce: '1731325253703',
-      permit: '0x',
-      partnerAndFee: '0',
-      bridge: {
-        protocolSelector: '0x00000000',
-        destinationChainId: 0,
-        outputToken: ZERO_ADDRESS,
-        scalingFactor: 0,
-        protocolData: '0x',
-      },
+    const { getDeltaPrice } = constructGetDeltaPrice({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    const price = await getDeltaPrice({
+      srcToken: WETH,
+      destToken: USDC_ARB,
+      amount: '1000000000000000000',
+      srcDecimals: 18,
+      destDecimals: 6,
+      destChainId: 42161,
+    });
+
+    expect(price.route.bridge?.protocol).toBe('Across');
+    expect(price.route.destination.input.token.chainId).toBe(42161);
+  });
+
+  test('getBridgeRoutes hits /delta/v2/prices/bridge-routes and unwraps `routes`', async () => {
+    const routes: BridgeRoute[] = [
+      { srcChainId: 1, destChainId: 42161, tokens: [USDC_ARB] },
+      { srcChainId: 1, destChainId: 10, tokens: [DAI] },
+    ];
+    const fetcher = makeFetcher(({ url, method }) => {
+      expect(method).toBe('GET');
+      expect(url).toBe(`${API_URL}/delta/v2/prices/bridge-routes`);
+      return { routes };
+    });
+
+    const { getBridgeRoutes } = constructGetBridgeRoutes({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    expect(await getBridgeRoutes()).toEqual(routes);
+  });
+
+  test('getBridgeRoutes passes filter params', async () => {
+    const fetcher = makeFetcher(({ url }) => {
+      expect(url).toContain('allowBridgeAndSwap=false');
+      expect(url).toContain('bridges=Across%2CRelay');
+      return { routes: [] };
+    });
+
+    const { getBridgeRoutes } = constructGetBridgeRoutes({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    expect(
+      await getBridgeRoutes({
+        allowBridgeAndSwap: false,
+        bridges: ['Across', 'Relay'],
+      })
+    ).toEqual([]);
+  });
+
+  test('isTokenSupportedInDelta unwraps `supported`', async () => {
+    const fetcher = makeFetcher(({ url, method }) => {
+      expect(method).toBe('GET');
+      expect(
+        url.startsWith(`${API_URL}/delta/v2/prices/is-token-supported/?`)
+      ).toBe(true);
+      expect(url).toContain(`token=${WETH}`);
+      expect(url).toContain('chainId=1');
+      return { supported: true };
+    });
+
+    const { isTokenSupportedInDelta } = constructIsTokenSupportedInDelta({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    expect(await isTokenSupportedInDelta(WETH)).toBe(true);
+  });
+
+  test('getDeltaOrders returns the pagination envelope', async () => {
+    const order = { id: 'auction-1' } as unknown as DeltaAuction;
+    const envelope: PaginatedResponse<DeltaAuction> = {
+      data: [order],
+      total: 1,
+      page: 1,
+      limit: 100,
+      hasMore: false,
     };
 
-    test('hash Delta Order', async () => {
-      const orderHash = await sdk.hashDeltaOrder(sampleOrder);
-      expect(orderHash).toMatchInlineSnapshot(
-        `"0x4e68492d838e64c329ecdba51d32cb088088445ca62b1d5c4edf5a2ab80b586d"`
+    const fetcher = makeFetcher(({ url, method }) => {
+      expect(method).toBe('GET');
+      expect(url.startsWith(`${API_URL}/delta/v2/orders?`)).toBe(true);
+      expect(url).toContain(`userAddress=${OWNER}`);
+      expect(url).toContain('page=2');
+      expect(url).toContain('limit=10');
+      return envelope;
+    });
+
+    const { getDeltaOrders } = constructGetDeltaOrders({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    const result = await getDeltaOrders({
+      userAddress: OWNER,
+      page: 2,
+      limit: 10,
+    });
+
+    expect(result).toEqual(envelope);
+    expect(result.hasMore).toBe(false);
+    expect(result.data).toHaveLength(1);
+  });
+
+  test('getDeltaOrders by id / by hash use the v2 path', async () => {
+    const order = { id: 'auction-1' } as unknown as DeltaAuction;
+    const fetcher = makeFetcher(({ url }) => {
+      if (url === `${API_URL}/delta/v2/orders/auction-1`) return order;
+      if (url === `${API_URL}/delta/v2/orders/hash/0xhash`) return order;
+      throw new Error(`unexpected URL ${url}`);
+    });
+
+    const { getDeltaOrderById, getDeltaOrderByHash } =
+      constructGetDeltaOrders({ apiURL: API_URL, chainId: 1, fetcher });
+
+    expect(await getDeltaOrderById('auction-1')).toBe(order);
+    expect(await getDeltaOrderByHash('0xhash')).toBe(order);
+  });
+
+  test('getRequiredBalanceForDeltaOrders hits /delta/v2/orders/fillablebalance/:chainId/:userAddress', async () => {
+    const required = {
+      [WETH]: '500000000000000000',
+      [DAI]: '1500000000000000000000',
+    };
+    const fetcher = makeFetcher(({ url, method }) => {
+      expect(method).toBe('GET');
+      expect(url).toBe(
+        `${API_URL}/delta/v2/orders/fillablebalance/1/${OWNER}`
       );
+      return required;
     });
-    test('PreSign Delta Order', async () => {
-      const sampleOrderHash =
-        '0x4e68492d838e64c329ecdba51d32cb088088445ca62b1d5c4edf5a2ab80b586d';
-      const txHash = await sdk.setDeltaOrderPreSignature(sampleOrderHash);
 
-      await viemWalletClient.waitForTransactionReceipt({ hash: txHash });
+    const { getRequiredBalanceForDeltaOrders } =
+      constructGetDeltaOrders({ apiURL: API_URL, chainId: 1, fetcher });
 
-      const isPreSigned = await checkIfOrderHashPreSigned({
-        orderHash: sampleOrderHash,
-        owner: senderAddress,
-        sdk,
-      });
-      expect(isPreSigned).toBe(true);
+    expect(
+      await getRequiredBalanceForDeltaOrders({ userAddress: OWNER })
+    ).toEqual(required);
+  });
+
+  test('getRequiredBalanceForDeltaOrders narrows the URL when tokenAddress is passed', async () => {
+    const fetcher = makeFetcher(({ url }) => {
+      expect(url).toBe(
+        `${API_URL}/delta/v2/orders/fillablebalance/1/${OWNER}/${WETH}`
+      );
+      return { [WETH]: '500000000000000000' };
     });
+
+    const { getRequiredBalanceForDeltaOrders } =
+      constructGetDeltaOrders({ apiURL: API_URL, chainId: 1, fetcher });
+
+    const balance = await getRequiredBalanceForDeltaOrders({
+      userAddress: OWNER,
+      tokenAddress: WETH,
+    });
+    expect(balance[WETH]).toBe('500000000000000000');
+  });
+
+  test('getAgentsList hits /delta/v2/agents/list/:chainId and returns agent names', async () => {
+    const agents = ['agent-a', 'agent-b'];
+    const fetcher = makeFetcher(({ url, method }) => {
+      expect(method).toBe('GET');
+      expect(url).toBe(`${API_URL}/delta/v2/agents/list/42161`);
+      return agents; // server returns string[] directly
+    });
+
+    const { getAgentsList } = constructGetAgentsList({
+      apiURL: API_URL,
+      chainId: 42161,
+      fetcher,
+    });
+
+    expect(await getAgentsList()).toEqual(agents);
   });
 });
 
-function getTokenAllowance({
-  tokenAddress,
-  owner,
-  spender,
-}: {
-  tokenAddress: string;
-  owner: string;
-  spender: string;
-}): Promise<BigNumber> {
-  const contract = new ethers.Contract(tokenAddress, erc20abi, signer);
-  return contract.allowance(owner, spender);
-}
+describe('Delta: build (server-side via POST /v2/orders/build)', () => {
+  test('buildDeltaOrder POSTs to /delta/v2/orders/build with correct body', async () => {
+    const builtFixture = buildBuiltOrderFixture();
+    let postedBody: any;
 
-function decreaseBySlippage(amount: string, slippagePercent: number): string {
-  const amountAfterSlippage = BigInt(
-    +(+amount * (1 - slippagePercent / 100)).toFixed(0)
-  ).toString(10);
+    const fetcher = makeFetcher(({ url, method, data }) => {
+      expect(method).toBe('POST');
+      expect(url).toBe(`${API_URL}/delta/v2/orders/build`);
+      postedBody = data;
+      return builtFixture;
+    });
 
-  return amountAfterSlippage;
-}
+    const { buildDeltaOrder } = constructBuildDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
 
-type VerifySignedCancelRequestInput = {
-  orderId: string;
-  signature: string;
-  address: string;
-  chainId: number;
-};
+    const route = buildPriceFixture().route;
+    const result = await buildDeltaOrder({
+      owner: OWNER,
+      route,
+      side: 'SELL',
+      partnerAddress: '0x0000000000000000000000000000000000000000',
+    });
 
-async function verifySignedCancelRequest({
-  orderId,
-  signature,
-  address,
-}: VerifySignedCancelRequestInput): Promise<boolean> {
-  const ParaswapDelta = await constructGetDeltaContract({
-    chainId,
-    fetcher: fetchFetcher,
-  }).getDeltaContract();
-
-  assert(ParaswapDelta, 'ParaswapDelta is not defined');
-
-  const valid = await verifyTypedData({
-    address: address as Address,
-    domain: {
-      name: 'Portikus',
-      version: '2.0.0',
-      chainId,
-      verifyingContract: ParaswapDelta as Address,
-    },
-    types: {
-      OrderCancellations: [{ name: 'orderIds', type: 'string[]' }],
-    },
-    primaryType: 'OrderCancellations',
-    message: {
-      orderIds: [orderId],
-    },
-    signature: signature as Hex,
+    expect(result).toEqual(builtFixture);
+    // build endpoint derives the chain from the route — no chainId in the body
+    expect(postedBody.chainId).toBeUndefined();
+    expect(postedBody.side).toBe('SELL');
+    expect(postedBody.owner).toBe(OWNER);
+    expect(postedBody.route).toBe(route);
+    expect(postedBody.slippage).toBeUndefined();
+    expect(postedBody.orderType).toBe('Order');
   });
 
-  return valid;
-}
+  test('buildDeltaOrder passes slippage to server for SELL', async () => {
+    const builtFixture = buildBuiltOrderFixture();
+    let postedBody: any;
 
-const PreSignatureModuleAbi = [
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'owner',
-        type: 'address',
-      },
-      {
-        internalType: 'bytes32',
-        name: 'orderHash',
-        type: 'bytes32',
-      },
-    ],
-    name: 'isPreSigned',
-    outputs: [
-      {
-        internalType: 'bool',
-        name: '',
-        type: 'bool',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'orderHash',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'bool',
-        name: 'preSigned',
-        type: 'bool',
-      },
-    ],
-    name: 'setPreSignature',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-] as const;
+    const fetcher = makeFetcher(({ url, data }) => {
+      if (url === `${API_URL}/delta/v2/orders/build`) {
+        postedBody = data;
+        return builtFixture;
+      }
+      throw new Error(`unexpected ${url}`);
+    });
 
-type CheckIfOrderHashPreSignedInput = {
-  orderHash: string;
-  owner: string;
-  sdk: GetDeltaContractFunctions;
-};
+    const { buildDeltaOrder } = constructBuildDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
 
-async function checkIfOrderHashPreSigned({
-  orderHash,
-  owner,
-  sdk,
-}: CheckIfOrderHashPreSignedInput): Promise<boolean> {
-  const ParaswapDelta = await sdk.getDeltaContract();
+    await buildDeltaOrder({
+      owner: OWNER,
+      route: buildPriceFixture().route,
+      side: 'SELL',
+      slippage: 100,
+      partnerAddress: '0x0000000000000000000000000000000000000000',
+    });
 
-  assert(ParaswapDelta, 'ParaswapDelta is not available');
-
-  const isPreSigned = await viemWalletClient.readContract({
-    address: ParaswapDelta as Address,
-    abi: PreSignatureModuleAbi,
-    functionName: 'isPreSigned',
-    args: [owner as Address, orderHash as Hash],
+    // slippage is forwarded to server; server applies it to destAmount
+    expect(postedBody.slippage).toBe(100);
+    expect(postedBody.side).toBe('SELL');
   });
-  return isPreSigned;
-}
+
+  test('buildDeltaOrder passes slippage to server for BUY', async () => {
+    const builtFixture = buildBuiltOrderFixture();
+    let postedBody: any;
+
+    const fetcher = makeFetcher(({ url, data }) => {
+      if (url === `${API_URL}/delta/v2/orders/build`) {
+        postedBody = data;
+        return builtFixture;
+      }
+      throw new Error(`unexpected ${url}`);
+    });
+
+    const { buildDeltaOrder } = constructBuildDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    await buildDeltaOrder({
+      owner: OWNER,
+      route: buildPriceFixture().route,
+      side: 'BUY',
+      slippage: 100,
+      partnerAddress: '0x0000000000000000000000000000000000000000',
+    });
+
+    expect(postedBody.slippage).toBe(100);
+    expect(postedBody.side).toBe('BUY');
+  });
+
+  test('buildDeltaOrder passes cross-chain route as-is; server injects destinationChainId', async () => {
+    const ccRoute = buildCrosschainRoute();
+    const builtFixture = buildBuiltOrderFixture({
+      srcToken: WETH,
+      destToken: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
+      bridge: {
+        protocolSelector: '0xdeadbeef',
+        destinationChainId: 42161,
+        outputToken: USDC_ARB,
+        scalingFactor: -12,
+        protocolData: '0xabcd',
+      },
+    });
+    let postedBody: any;
+
+    const fetcher = makeFetcher(({ url, data }) => {
+      if (url === `${API_URL}/delta/v2/orders/build`) {
+        postedBody = data;
+        return builtFixture;
+      }
+      throw new Error(`unexpected ${url}`);
+    });
+
+    const { buildDeltaOrder } = constructBuildDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    const result = await buildDeltaOrder({
+      owner: OWNER,
+      route: ccRoute,
+      side: 'SELL',
+      partnerAddress: '0x0000000000000000000000000000000000000000',
+    });
+
+    // route is forwarded as-is; bridge in route.contractParams has no destinationChainId
+    expect(postedBody.route).toBe(ccRoute);
+    // the server (fixture) fills in destinationChainId
+    expect(result.toSign.value.bridge).toMatchObject({
+      protocolSelector: '0xdeadbeef',
+      destinationChainId: 42161,
+    });
+  });
+
+  test('buildExternalDeltaOrder sends orderType ExternalOrder with handler/data', async () => {
+    const builtFixture = buildBuiltOrderFixture();
+    let postedBody: any;
+
+    const fetcher = makeFetcher(({ url, data }) => {
+      if (url === `${API_URL}/delta/v2/orders/build`) {
+        postedBody = data;
+        return builtFixture;
+      }
+      throw new Error(`unexpected ${url}`);
+    });
+
+    const { buildExternalDeltaOrder } = constructBuildExternalDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    const route = buildPriceFixture().route;
+    await buildExternalDeltaOrder({
+      owner: OWNER,
+      handler: '0x2222222222222222222222222222222222222222',
+      data: '0xdeadbeef',
+      route,
+      side: 'SELL',
+      partnerAddress: '0x0000000000000000000000000000000000000000',
+    });
+
+    expect(postedBody.orderType).toBe('ExternalOrder');
+    expect(postedBody.handler).toBe(
+      '0x2222222222222222222222222222222222222222'
+    );
+    expect(postedBody.data).toBe('0xdeadbeef');
+    expect(postedBody.route).toBe(route);
+  });
+
+  test('buildTWAPDeltaOrder (sell) sends TWAPOrder body; slippage forwarded to server', async () => {
+    const builtFixture = buildBuiltOrderFixture();
+    let postedBody: any;
+
+    const fetcher = makeFetcher(({ url, data }) => {
+      if (url === `${API_URL}/delta/v2/orders/build`) {
+        postedBody = data;
+        return builtFixture;
+      }
+      throw new Error(`unexpected ${url}`);
+    });
+
+    const { buildTWAPDeltaOrder } = constructBuildTWAPDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    const route = buildPriceFixture().route;
+    await buildTWAPDeltaOrder({
+      owner: OWNER,
+      onChainOrderType: 'TWAPOrder',
+      route,
+      totalSrcAmount: '5000000000000000000',
+      interval: 300,
+      numSlices: 5,
+      slippage: 50,
+      partnerAddress: '0x0000000000000000000000000000000000000000',
+    });
+
+    expect(postedBody.orderType).toBe('TWAPOrder');
+    expect(postedBody.interval).toBe(300);
+    expect(postedBody.numSlices).toBe(5);
+    expect(postedBody.totalSrcAmount).toBe('5000000000000000000');
+    expect(postedBody.slippage).toBe(50); // server applies slippage to destAmountPerSlice
+    expect(postedBody.side).toBe('SELL');
+  });
+
+  test('buildTWAPDeltaOrder (buy) forwards slippage and maxSrcAmount to server', async () => {
+    const builtFixture = buildBuiltOrderFixture();
+    let postedBody: any;
+
+    const fetcher = makeFetcher(({ url, data }) => {
+      if (url === `${API_URL}/delta/v2/orders/build`) {
+        postedBody = data;
+        return builtFixture;
+      }
+      throw new Error(`unexpected ${url}`);
+    });
+
+    const { buildTWAPDeltaOrder } = constructBuildTWAPDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    const route = buildPriceFixture().route;
+    await buildTWAPDeltaOrder({
+      owner: OWNER,
+      onChainOrderType: 'TWAPBuyOrder',
+      route,
+      totalDestAmount: '5000000000000000000',
+      maxSrcAmount: '1000000000000000000',
+      interval: 300,
+      numSlices: 5,
+      slippage: 100,
+      partnerAddress: '0x0000000000000000000000000000000000000000',
+    });
+
+    expect(postedBody.orderType).toBe('TWAPBuyOrder');
+    expect(postedBody.side).toBe('BUY');
+    expect(postedBody.maxSrcAmount).toBe('1000000000000000000');
+    expect(postedBody.slippage).toBe(100);
+  });
+});
+
+describe('Delta: submit (build → sign → post)', () => {
+  test('submitDeltaOrder posts to /delta/v2/orders with signed order', async () => {
+    const builtFixture = buildBuiltOrderFixture();
+    let posted: any;
+    let postUrl = '';
+
+    const fetcher = makeFetcher(({ url, method, data }) => {
+      if (method === 'POST' && url === `${API_URL}/delta/v2/orders/build`) {
+        return builtFixture;
+      }
+      if (method === 'POST' && url.startsWith(`${API_URL}/delta/v2/orders`)) {
+        postUrl = url;
+        posted = data;
+        return {
+          id: 'auction-99',
+          order: data.order,
+          onChainOrderType: 'Order',
+        } as unknown as DeltaAuction<'Order'>;
+      }
+      throw new Error(`unexpected request ${method} ${url}`);
+    });
+
+    const { submitDeltaOrder } = constructSubmitDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+      contractCaller: makeMockContractCaller(),
+    });
+
+    const response = await submitDeltaOrder({
+      owner: OWNER,
+      route: buildPriceFixture().route,
+      side: 'SELL',
+      partnerAddress: '0x0000000000000000000000000000000000000000',
+    });
+
+    expect(postUrl.startsWith(`${API_URL}/delta/v2/orders`)).toBe(true);
+    expect(posted.signature).toBe(FAKE_SIGNATURE);
+    expect(posted.chainId).toBe(1);
+    // order comes from builtFixture.toSign.value
+    expect(posted.order.owner).toBe(OWNER);
+    expect(posted.order.srcToken).toBe(WETH);
+    expect(posted.order.destToken).toBe(DAI);
+    expect(response.id).toBe('auction-99');
+  });
+
+  test('postDeltaOrder forwards degenMode as a query param', async () => {
+    const fetcher = makeFetcher(({ url, method }) => {
+      expect(method).toBe('POST');
+      expect(url).toContain('degenMode=true');
+      return { id: 'x' } as unknown as DeltaAuction<'Order'>;
+    });
+
+    const { postDeltaOrder } = constructPostDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    expect(
+      await postDeltaOrder({
+        signature: FAKE_SIGNATURE,
+        order: {} as any,
+        degenMode: true,
+      })
+    ).toEqual({ id: 'x' });
+  });
+
+  test('postExternalDeltaOrder sends to /delta/v2/orders', async () => {
+    const fetcher = makeFetcher(({ url, method }) => {
+      expect(method).toBe('POST');
+      expect(url).toBe(`${API_URL}/delta/v2/orders`);
+      return { id: 'ext-1' } as unknown as DeltaAuction<'ExternalOrder'>;
+    });
+
+    const { postExternalDeltaOrder } = constructPostExternalDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    expect(
+      await postExternalDeltaOrder({
+        signature: FAKE_SIGNATURE,
+        order: {} as any,
+      })
+    ).toEqual({ id: 'ext-1' });
+  });
+
+  test('postTWAPDeltaOrder sends to /delta/v2/orders with onChainOrderType', async () => {
+    const fetcher = makeFetcher(({ url, method, data }) => {
+      expect(method).toBe('POST');
+      expect(url.startsWith(`${API_URL}/delta/v2/orders`)).toBe(true);
+      expect(data.onChainOrderType).toBe('TWAPOrder');
+      return { id: 'twap-1' } as unknown as DeltaAuction<'TWAPOrder'>;
+    });
+
+    const { postTWAPDeltaOrder } = constructPostTWAPDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+    });
+
+    expect(
+      await postTWAPDeltaOrder({
+        signature: FAKE_SIGNATURE,
+        order: {} as any,
+        onChainOrderType: 'TWAPOrder',
+      })
+    ).toEqual({ id: 'twap-1' });
+  });
+});
+
+describe('Delta: cancel', () => {
+  test('cancelDeltaOrders signs then posts to /delta/v2/orders/cancel', async () => {
+    let postedTo = '';
+    let postedBody: any;
+
+    const fetcher = makeFetcher(({ url, method, data }) => {
+      const adapter = url.includes('/adapters/contracts')
+        ? {
+            AugustusSwapper: '0x',
+            TokenTransferProxy: '0x',
+            AugustusRFQ: '0x',
+            Executors: {},
+            ParaswapDelta: '0x1111111111111111111111111111111111111111',
+          }
+        : undefined;
+      if (adapter) return adapter;
+      if (method === 'POST') {
+        postedTo = url;
+        postedBody = data;
+        return { success: true };
+      }
+      throw new Error('unexpected');
+    });
+
+    const { cancelDeltaOrders } = constructCancelDeltaOrder({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher,
+      contractCaller: makeMockContractCaller(),
+    });
+
+    const result = await cancelDeltaOrders({
+      orderIds: ['a', 'b'],
+    });
+
+    expect(result).toEqual({ success: true });
+    expect(postedTo).toBe(`${API_URL}/delta/v2/orders/cancel`);
+    expect(postedBody.orderIds).toEqual(['a', 'b']);
+    expect(postedBody.signature).toBe(FAKE_SIGNATURE);
+  });
+});
+
+describe('Delta: live API contract', () => {
+  jest.setTimeout(30_000);
+
+  const LIVE_API = process.env.API_URL;
+  const fetchFetcher = constructFetchFetcher(fetch);
+
+  test('GET /delta/v2/prices (same-chain) matches DeltaPrice shape', async () => {
+    const { getDeltaPrice } = constructGetDeltaPrice({
+      apiURL: LIVE_API,
+      chainId: 1,
+      fetcher: fetchFetcher,
+    });
+
+    const price = await getDeltaPrice({
+      srcToken: WETH,
+      destToken: DAI,
+      amount: '1000000000000000000',
+      srcDecimals: 18,
+      destDecimals: 18,
+      side: 'SELL',
+    });
+
+    expect(price.side).toBe('SELL');
+    expect(price.inputToken.chainId).toBe(1);
+    expect(price.outputToken.chainId).toBe(1);
+    expect(typeof price.id).toBe('string');
+    expect(typeof price.spender).toBe('string');
+    expect(price.route.origin.input.token.address.toLowerCase()).toBe(WETH);
+    expect(typeof price.route.origin.input.amount).toBe('string');
+    expect(typeof price.route.origin.input.amountUSD).toBe('string');
+    // same-chain: bridge is null
+    expect(price.route.bridge).toBeNull();
+    // fees.gas is a single DeltaTokenAmount, fees.bridge is an array
+    expect(typeof price.route.fees.gas.amount).toBe('string');
+    expect(Array.isArray(price.route.fees.bridge)).toBe(true);
+  });
+
+  test('GET /delta/v2/prices (cross-chain) returns DeltaRoute with bridge.contractParams (no destinationChainId)', async () => {
+    const { getDeltaPrice } = constructGetDeltaPrice({
+      apiURL: LIVE_API,
+      chainId: 1,
+      fetcher: fetchFetcher,
+    });
+
+    const price = await getDeltaPrice({
+      srcToken: WETH,
+      destToken: USDC_ARB,
+      amount: '1000000000000000000',
+      srcDecimals: 18,
+      destDecimals: 6,
+      destChainId: 42161,
+      side: 'SELL',
+    });
+
+    expect(price.outputToken.chainId).toBe(42161);
+    expect(price.route.destination.input.token.chainId).toBe(42161);
+    expect(price.route.bridge).not.toBeNull();
+    if (price.route.bridge) {
+      expect(typeof price.route.bridge.protocol).toBe('string');
+      expect(Array.isArray(price.route.bridge.tags)).toBe(true);
+      expect(typeof price.route.bridge.contractParams.protocolSelector).toBe(
+        'string'
+      );
+      expect(typeof price.route.bridge.contractParams.outputToken).toBe(
+        'string'
+      );
+      // destinationChainId is NOT present on the wire — the SDK injects it at build time
+      expect('destinationChainId' in price.route.bridge.contractParams).toBe(
+        false
+      );
+    }
+  });
+
+  test('GET /delta/v2/prices/bridge-routes returns flat array', async () => {
+    const { getBridgeRoutes } = constructGetBridgeRoutes({
+      apiURL: LIVE_API,
+      chainId: 1,
+      fetcher: fetchFetcher,
+    });
+
+    const routes = await getBridgeRoutes();
+    expect(Array.isArray(routes)).toBe(true);
+    expect(routes.length).toBeGreaterThan(0);
+    const first = routes[0]!;
+    expect(typeof first.srcChainId).toBe('number');
+    expect(typeof first.destChainId).toBe('number');
+    expect(Array.isArray(first.tokens)).toBe(true);
+  });
+
+  test('GET /delta/v2/prices/bridge-protocols returns protocols', async () => {
+    const { getBridgeProtocols } = constructGetBridgeRoutes({
+      apiURL: LIVE_API,
+      chainId: 1,
+      fetcher: fetchFetcher,
+    });
+
+    const protocols = await getBridgeProtocols();
+    expect(protocols.length).toBeGreaterThan(0);
+    expect(protocols.some((p) => p.protocol === 'Across')).toBe(true);
+  });
+});
+
+describe('Delta: SDK wiring', () => {
+  test('constructAllDeltaOrdersHandlers exposes all v2 methods', () => {
+    const sdk = constructAllDeltaOrdersHandlers({
+      apiURL: API_URL,
+      chainId: 1,
+      fetcher: makeFetcher(() => ({})),
+      contractCaller: makeMockContractCaller(),
+    });
+
+    expect(typeof sdk.getDeltaPrice).toBe('function');
+    expect(typeof sdk.getDeltaOrders).toBe('function');
+    expect(typeof sdk.getRequiredBalanceForDeltaOrders).toBe('function');
+    expect(typeof sdk.getBridgeRoutes).toBe('function');
+    expect(typeof sdk.buildDeltaOrder).toBe('function');
+    expect(typeof sdk.postDeltaOrder).toBe('function');
+    expect(typeof sdk.submitDeltaOrder).toBe('function');
+    expect(typeof sdk.submitExternalDeltaOrder).toBe('function');
+    expect(typeof sdk.submitTWAPDeltaOrder).toBe('function');
+    expect(typeof sdk.cancelDeltaOrders).toBe('function');
+    expect(typeof sdk.isTokenSupportedInDelta).toBe('function');
+    expect(typeof sdk.getAgentsList).toBe('function');
+    // reused v1 utilities
+    expect(typeof sdk.getDeltaContract).toBe('function');
+    expect(typeof sdk.getPartnerFee).toBe('function');
+    expect(typeof sdk.approveTokenForDelta).toBe('function');
+    // v2 sign function (replaces v1 sign in deltaV2 namespace)
+    expect(typeof sdk.signDeltaOrder).toBe('function');
+  });
+
+  test('constructPartialSDK accepts v2 constructors individually', () => {
+    const fetcher = makeFetcher(() => ({}));
+    const sdk = constructPartialSDK(
+      { apiURL: API_URL, chainId: 1, fetcher },
+      constructBuildDeltaOrder,
+      constructPostDeltaOrder,
+      constructGetDeltaOrders,
+      constructGetDeltaPrice,
+      constructGetDeltaContract,
+      constructGetPartnerFee,
+      constructGetBridgeRoutes,
+      constructIsTokenSupportedInDelta
+    );
+
+    expect(typeof sdk.getDeltaPrice).toBe('function');
+    expect(typeof sdk.getBridgeRoutes).toBe('function');
+    expect(typeof sdk.getDeltaOrders).toBe('function');
+    expect(typeof sdk.getRequiredBalanceForDeltaOrders).toBe('function');
+    expect(typeof sdk.buildDeltaOrder).toBe('function');
+    expect(typeof sdk.postDeltaOrder).toBe('function');
+    expect(typeof sdk.isTokenSupportedInDelta).toBe('function');
+  });
+});
